Feature/counterdown screen

+++ feature/countdown/src/main/java/com/jetpack/countdown/model/CountdownState.kt
@@ -4,5 +4,5 @@ package com.jetpack.countdown.model
 data class CountdownState(
     val counterState: CounterState = CounterState.INITIAL,
     val workingState: WorkingState = WorkingState.WORK,
-    val remainTime: Long = WORKING_DURATION,
+    val remainTime: Int = WORKING_DURATION,
 )
